[FTR] Added scheduler subsystem
Added the Scheduler subsystem, a system for interfacing with the
operating systems clock and managing the delivery of updates to classes
that the user defines.

This system is fairly simple and largely feature complete but may
receive optimisations and minor changes in the distant future (E.G:
Greater flexibility in scheduled item update skipping).